Issue #12012: test_ssl uses test_support.import_module()
Skip the whole file if the SSL module is missing. It was already the case,
except that the SkipTest exception was raised in test_main().

This commit fixes an error in test_ssl if the ssl module is missing.

--HG--
branch : 2.7
extra : rebase_source : a952be35b1f368b3e3731a5714d3db7682e51932