Make the from_iterable() recipe more usable.
The code isn't exactly equivalent because a classmethod would
only make sense inside a chain class, and it would need "cls"
as a first argument, and it would need to return an instance
of "chain" rather than a generator.

The updated example drops the @classmethod decorator so that
it can be used standalone:  list(from_iterable(['abc', 'def']))
This should be communicate what from_iterable does.

--HG--
branch : 2.7