#15545: sort iterdump via SQL instead of in python code
Although there is not a regression in Python2, we make the same update here to
keep the code bases in sync.

(The fix for issue 9750 introduced a regression in Python 3 by sorting the row
objects returned by fetchall.  But if a row_factory such as sqlite3.Row is
used, the rows may not be sortable (in Python3), which leads to an exception.
The sorting is still a nice idea, so the patch moves the sort into the sql.)

Fix and test by Peter Otten.

--HG--
branch : 2.7