Issue #18609, #18408: _ctypes_add_traceback() now clears the current exception while adding the traceback, because it may call indirectly a Python function and Python functions must not be called with an exception set.
In the case of the issue #18609, _ctypes_add_traceback() called the iso8859-1
decoder which is implemented in Python. Python has a ISO-8859-1 codec
implemented in C. It is not used because PyUnicode_Decode() only uses the C
codec for other names (aliases) of this codec ("latin-1", "latin1" and
"iso-8859-1").

--HG--
extra : rebase_source : 55cc60b7bd05bcaddddb67c46c9c19369d793803