#20135: FAQ entry for list mutation.  (See also 6375bf34fff6.)
I accidentally committed this in 6375bf34fff6; this changeset addresses the
last review comments on the patch.

This is a perennial question and something someone opens a ticket for probably
every other month or so, so I'm surprised we didn't already have a FAQ entry
for it.

The original patch was written by M. Votz, refined first by Ezio Melotti and
further refined by me.

--HG--
branch : 3.4