#12780: update inspect test skipIf for PEP 3147.
The test needs to be skipped if unicodedata is either part of the
main binary (a repackaging of cpython on Windows?) or has python
source (pypy?).  PEP 3147 makes __file__ point to the .py source,
so we need to change the extension check from looking for the
old .pyc/.pyo to just looking for .py.

Note that this skip should never trigger on CPython itself, so
one could argue it should be dropped instead.  But since it exists,
why risk breaking someone else's python.

--HG--
branch : 3.4