Issue #23485: Enhance and update selectors doc and test_selectors
Selector.select() is now retried with the recomputed timeout when interrupted
by a signal.

Write an unit test with a signal handler raising an exception, and a unit with
a signal handler which does not raise an exception (it does nothing).

--HG--
extra : histedit_source : 4c78495fd6e9d703967e7fb74fb70f36dc1e4810