#11754: test contents of string module attributes.
As noted in the comment, while the order of the items in the attributes is not
technically guaranteed, after all this time there is almost certainly user
code out there that relies on it, so we might as well test for it.

Patch by Chalmer Lowe.

--HG--
extra : amend_source : 242d5531e4f953256b6afb332c337d3807ec3b71