Rebase the 3.x version of idlelib/help.html on the 3.5 version of Doc/build/html/library/idle.html, which is built from Doc/library/idle.rst. Help.html is an rstrip()ed copy of idle.html that can be pushed.
The displayed content is unchanged because a) the version number is not
displayed, being inherent in the version displaying the file and b) the tt
tag was ignored by the custom parser in idlelib/help.py.  The wrapped
<span class="pre"> tags remain as they were.

This will be merged in 3.6 (without conflict) because the 3.6 help.html is
a copy of the 3.5 file (which was a copy of the 3.4 file).  The two files will
remain the same until either a) there is a content change that only applies
to one of them or b) 3.5 maintenance ends and the 3.6 file is itself rebased.

--HG--
branch : 3.5