#24277: The new email API is no longer provisional.
This is a wholesale reorganization and editing of the email documentation to
make the new API the standard one, and the old API the 'legacy' one.  The
default is still the compat32 policy, for backward compatibility.  We will
change that eventually.

--HG--
rename : Doc/includes/email-read-alternative-new-api.py => Doc/includes/email-read-alternative.py
rename : Doc/library/email-examples.rst => Doc/library/email.examples.rst